Only check direct parent for ignore/mark flags
The flags of a new message should be based on the flags from the direct
parent, not from any other parent in the "In-Reply-To" chain.

Consider a thread like this:
Msg-ID: A
 Msg-ID: B; In-Reply-To: A
  Msg-ID: C; In-Reply-To: B
   Msg-ID: D; In-Reply-To: C

If the messages A and/or B have a flag, and C does not, the current
code walks up to B to decide the flags for D. It is legitimate for C to
have different flags than its parents. This could happen if someone
hijacks an already ignored thread A, with the new topic starting at C.

Signed-off-by: Olaf Hering <olaf@aepfle.de>